+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- [val-p7m] P7MContentValidator (validate p7m and possibly its content with a facade)
 - [val-pdfbox] PdfboxStrictValidator (uses not lenient parser)
 
 ### Changed